design/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ectf25_design"
version = "1.0.0a6"
description = "Hammer Industries' (Purdue2) encoder design for the MITRE eCTF 2025."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Noah Rediker", email = "nrediker@purdue.edu"},
    {name = "Spencer Bowles", email = "bowles7@purdue.edu"},
    {name = "Aidan Jacobsen", email = "jacobse7@purdue.edu"},
    {name = "Eli Wood", email = "wood269@purdue.edu"}
]
keywords = ["mitre", "ectf", "embedded", "security", "encoder", "aes", "purdue"]
dependencies = ["pycryptodome>=3.21"]

[project.optional-dependencies]
tests = ["pythonfuzz"]

[tool.setuptools.packages.find]
exclude = ["tests*"]
